adding routing for event's dynamic page
